+++ term_image/tui/widgets.py
@@ -6,15 +6,15 @@ import logging as _logging
 from math import ceil
 from operator import floordiv, mul, sub
 from os.path import basename
-from shutil import get_terminal_size
 from typing import List, Optional, Tuple
 
 import urwid
 
 from .. import logging
 from ..config import _nav, cell_width, expand_key, nav
-from ..image import TermImage
+from ..image import BaseImage
 from ..image.common import _ALPHA_THRESHOLD
+from ..utils import get_terminal_size
 from . import keys, main as tui_main
 from .render import grid_render_queue, image_render_queue
 
@@ -170,7 +170,7 @@ class Image(urwid.Widget):
 
     _alpha = f"{_ALPHA_THRESHOLD}"[1:]  # Updated from `.tui.init()`
 
-    def __init__(self, image: TermImage):
+    def __init__(self, image: BaseImage):
         self._image = image
 
     def keypress(self, size: Tuple[int, int], key: str) -> str: